// src/data/types.ts
export interface UnitRow {
  Id: number
  Name: string | null
  HUDName: string | null
  Description: string | null
  CountryId: number
  Cost: number
  Type: number
  CategoryType: number
  Role: number
  Length: number
  Width: number
  Height: number
  Weight: number
  Stealth: number
  InfantrySlots: number
  MaxStress: number
  /** seconds added to the base "return to battlegroup delay" (0 for most units) */
  TimeModifierRefund: number
  ContentMembership: number
  DisplayInArmory: boolean
  IsUnitModification: boolean
  ModelFileName: string | null
  PortraitFileName: string | null
  ThumbnailFileName: string | null
  OriginalName: string | null
  OriginalCost: number
  OwnerInfantryID: number
}